Update DMPHN_1_2_4_8.py
Thanx for sharing the code, which help me understand deeply in mind, while I'm confusing about this part of code, where other parts also have this issue. As Your paper illustrated the level 3 of feature of Net after concatenated, which is in LEFT and RIGHT way, specifically, Not in TOP and BOTTOM way.